+++ Documentation/CHANGES.md
@@ -8,6 +8,10 @@ Do not try nor report any issue until a release is tagged Stable.
 
 ## Changes
 
+### Changes of the Release Beta 3.0.0
+
+- Pass.app: release 0.2. Added Edit button and action to the List Panel.
+
 ### Changes of the Release Candidate 2.0.4
 
 - Tweaking alsa test on pi 500.